+++ src/project/api/dish_products_routes.py
@@ -1,14 +1,19 @@
-from fastapi import APIRouter, HTTPException, status
+from fastapi import APIRouter, HTTPException, status, Depends
 
-from project.schemas.user import DishProductsCreate, DishProductsSchema
+from project.schemas.user import DishProductsCreate, DishProductsSchema, ClientSchema
 from project.core.exceptions import DishProductNotFound, DishProductAlreadyExists
-from project.api.depends import database, dishProducts_repo
+from project.api.depends import database, dishProducts_repo, get_current_client, check_for_admin_access
 
 
 dish_products_router = APIRouter()
 
 
-@dish_products_router.get("/all_dish_products", response_model=list[DishProductsSchema], status_code=status.HTTP_200_OK)
+@dish_products_router.get(
+    "/all_dish_products",
+    response_model=list[DishProductsSchema],
+    status_code=status.HTTP_200_OK,
+    dependencies = [Depends(get_current_client)],
+)
 async def get_all_dish_products() -> list[DishProductsSchema]:
     async with database.session() as session:
         all_dish_products = await dishProducts_repo.get_all_dish_products(session=session)
@@ -16,7 +21,12 @@ async def get_all_dish_products() -> list[DishProductsSchema]:
     return all_dish_products
 
 
-@dish_products_router.get("/dish_product/{dish_id}/{product_id}", response_model=DishProductsSchema, status_code=status.HTTP_200_OK)
+@dish_products_router.get(
+    "/dish_product/{dish_id}/{product_id}",
+    response_model=DishProductsSchema,
+    status_code=status.HTTP_200_OK,
+    dependencies=[Depends(get_current_client)],
+)
 async def get_dish_product(
     dish_id: int,
     product_id: int,
@@ -34,7 +44,12 @@ async def get_dish_product(
     return dish_product
 
 
-@dish_products_router.get("/dish/{dish_id}/products", response_model=list[DishProductsSchema], status_code=status.HTTP_200_OK)
+@dish_products_router.get(
+    "/dish/{dish_id}/products",
+    response_model=list[DishProductsSchema],
+    status_code=status.HTTP_200_OK,
+    dependencies=[Depends(get_current_client)],
+)
 async def get_products_for_dish(
     dish_id: int,
 ) -> list[DishProductsSchema]:
@@ -50,7 +65,9 @@ async def get_products_for_dish(
 @dish_products_router.post("/add_dish_product", response_model=DishProductsSchema, status_code=status.HTTP_201_CREATED)
 async def add_dish_product(
     dish_product_dto: DishProductsCreate,
+    current_client: ClientSchema = Depends(get_current_client),
 ) -> DishProductsSchema:
+    check_for_admin_access(client=current_client)
     try:
         async with database.session() as session:
             new_dish_product = await dishProducts_repo.create_dish_product(
@@ -72,7 +89,9 @@ async def update_dish_product(
     dish_id: int,
     product_id: int,
     dish_product_dto: DishProductsCreate,
+    current_client: ClientSchema = Depends(get_current_client),
 ) -> DishProductsSchema:
+    check_for_admin_access(client=current_client)
     try:
         async with database.session() as session:
             updated_dish_product = await dishProducts_repo.update_dish_product(
@@ -91,7 +110,9 @@ async def update_dish_product(
 async def delete_dish_product(
     dish_id: int,
     product_id: int,
+    current_client: ClientSchema = Depends(get_current_client),
 ) -> None:
+    check_for_admin_access(client=current_client)
     try:
         async with database.session() as session:
             await dishProducts_repo.delete_dish_product(
@@ -107,7 +128,9 @@ async def delete_dish_product(
 @dish_products_router.delete("/dish/{dish_id}/products", status_code=status.HTTP_204_NO_CONTENT)
 async def delete_all_products_for_dish(
     dish_id: int,
+    current_client: ClientSchema = Depends(get_current_client),
 ) -> None:
+    check_for_admin_access(client=current_client)
     async with database.session() as session:
         await dishProducts_repo.delete_all_products_for_dish(
             session=session,
