+++ src/project/api/table_routes.py
@@ -1,8 +1,8 @@
-from fastapi import APIRouter, HTTPException, status
+from fastapi import APIRouter, HTTPException, status, Depends
 
-from project.schemas.user import TableCreate, TableSchema
+from project.schemas.user import TableCreate, TableSchema, ClientSchema
 from project.core.exceptions import TableNotFound, TableAlreadyExists
-from project.api.depends import database, table_repo
+from project.api.depends import database, table_repo, check_for_admin_access, get_current_client
 
 
 table_router = APIRouter()
@@ -17,7 +17,12 @@ async def get_all_tables() -> list[TableSchema]:
     return all_tables
 
 
-@table_router.get("/table/{table_id}", response_model=TableSchema, status_code=status.HTTP_200_OK)
+@table_router.get(
+    "/table/{table_id}",
+    response_model=TableSchema,
+    status_code=status.HTTP_200_OK,
+    dependencies=[Depends(get_current_client)],
+)
 async def get_table_by_id(
     table_id: int,
 ) -> TableSchema:
@@ -33,7 +38,9 @@ async def get_table_by_id(
 @table_router.post("/add_table", response_model=TableSchema, status_code=status.HTTP_201_CREATED)
 async def add_table(
     table_dto: TableCreate,
+        current_client: ClientSchema = Depends(get_current_client),
 ) -> TableSchema:
+    check_for_admin_access(client=current_client)
     try:
         async with database.session() as session:
             new_table = await table_repo.create_table(session=session, table=table_dto)
@@ -68,7 +75,9 @@ async def update_table(
 @table_router.delete("/delete_table/{table_id}", status_code=status.HTTP_204_NO_CONTENT)
 async def delete_table(
     table_id: int,
+        current_client: ClientSchema = Depends(get_current_client),
 ) -> None:
+    check_for_admin_access(client=current_client)
     try:
         async with database.session() as session:
             table = await table_repo.delete_table(session=session, table_id=table_id)
